Refactor code to include components and allow for additional dashboard pages
