+++ README.md
@@ -1,0 +1,20 @@
+
+## How to Run
+
+1. Clone the repository
+
+```
+git clone https://github.com/dkimhw/dealership_competitor_dash.git
+```
+
+2. Install dependencies
+
+```
+pip install -r requirements.txt
+```
+
+3. Run the application
+
+```
+python index.py
+```
