+++ notes/code_to_check_data
